items: implement pool structure
