+++ BlasII.Randomizer/Items/Shuffle/ItemShuffler.cs
@@ -1,13 +1,17 @@
 ﻿using System.Collections.Generic;
 
-namespace BlasII.Randomizer.Items
+namespace BlasII.Randomizer.Items.Shuffle
 {
-    internal class ItemShufflerReverse : BaseShuffler
+    internal class ItemShuffler : BaseShuffler
     {
         public override bool Shuffle(int seed, RandomizerSettings settings, Dictionary<string, string> output)
         {
             output.Clear();
             Initialize(seed);
+
+            // Create inventory with starting items
+            var inventory = new Blas2Inventory(settings, Main.Randomizer.Data.DoorDictionary);
+            AddStartingItemsToInventory(inventory, settings);
 
             // Create pools of all locations to randomize
             List<ItemLocation> progressionLocations = new(), junkLocations = new();
@@ -17,26 +21,29 @@ namespace BlasII.Randomizer.Items
             List<Item> progressionItems = new(), junkItems = new();
             CreateItemPool(progressionItems, junkItems, progressionLocations.Count + junkLocations.Count, settings);
 
-            // Create initial inventory
-            var inventory = new Blas2Inventory(settings, Main.Randomizer.Data.DoorDictionary);
-            CreateInitialInventory(inventory, settings, progressionItems);
-
             // Place progression items at progression locations
             FillProgressionItems(progressionLocations, progressionItems, output, inventory);
 
-            // Add junk items to remaining progression items and ensure they are still balanced
-            junkLocations.AddRange(progressionLocations);
-            if (junkLocations.Count != junkItems.Count)
+            // Verify that all progression items were placed
+            if (progressionItems.Count > 0)
                 return false;
 
             // Place junk items at junk locations and remaining progression locations
+            junkLocations.AddRange(progressionLocations);
             FillJunkItems(junkLocations, junkItems, output);
 
             // Verify that all remaining items were placed
-            return junkItems.Count == 0 && junkLocations.Count == 0;
+            return junkItems.Count == 0;
         }
 
-        #region Setup
+        /// <summary>
+        /// Adds the starting weapon to the initial inventory
+        /// </summary>
+        private void AddStartingItemsToInventory(Blas2Inventory inventory, RandomizerSettings settings)
+        {
+            // Add the starting weapon
+            inventory.AddItem(Main.Randomizer.Data.GetItem(GetStartingWeaponId(settings)));
+        }
 
         /// <summary>
         /// Fills the two location pools
@@ -112,18 +119,27 @@ namespace BlasII.Randomizer.Items
         }
 
         /// <summary>
-        /// Adds the starting weapon and all progression items to the initial inventory
+        /// Calculates a subset of the given locations that are reachable with the current inventory
         /// </summary>
-        private void CreateInitialInventory(Blas2Inventory inventory, RandomizerSettings settings, List<Item> progressionItems)
+        private List<ItemLocation> FindReachableLocations(List<ItemLocation> locations, Blas2Inventory inventory)
         {
-            // Add the starting weapon
-            inventory.AddItem(Main.Randomizer.Data.GetItem(GetStartingWeaponId(settings)));
-
-            // Add all progression items in the pool
-            foreach (var item in progressionItems)
+            var reachableLocations = new List<ItemLocation>();
+            foreach (var location in locations)
             {
-                inventory.AddItem(item);
+                if (inventory.Evaluate(location.logic))
+                    reachableLocations.Add(location);
             }
+            return reachableLocations;
+        }
+
+        /// <summary>
+        /// After shuffling the list of progression items, move certain items to the end to prevent failing seeds
+        /// </summary>        
+        private void MovePriorityItems(List<Item> progressionItems)
+        {
+            Item wallClimb = Main.Randomizer.Data.GetItem("AB44");
+            progressionItems.Remove(wallClimb);
+            progressionItems.Add(wallClimb);
         }
 
         /// <summary>
@@ -134,44 +150,27 @@ namespace BlasII.Randomizer.Items
             return "WE0" + (settings.RealStartingWeapon + 1);
         }
 
-        #endregion
 
-        #region Shuffle
 
-        /// <summary>
-        /// Using a reverse fill algorithm, place the last item at a random reachable location
-        /// </summary>
         private void FillProgressionItems(List<ItemLocation> locations, List<Item> items, Dictionary<string, string> output, Blas2Inventory inventory)
         {
-            // Verify that all locations are reachable
-            foreach (var location in locations)
-            {
-                if (!inventory.Evaluate(location.logic))
-                    return;
-            }
-
             ShuffleList(items);
             MovePriorityItems(items);
-            var reachableLocations = new List<ItemLocation>(locations);
+            List<ItemLocation> reachableLocations = FindReachableLocations(locations, inventory);
 
             while (reachableLocations.Count > 0 && items.Count > 0)
             {
+                ItemLocation location = RemoveRandomFromOther(reachableLocations, locations);
                 Item item = RemoveLast(items);
-                inventory.RemoveItem(item);
 
-                RemoveUnreachableLocations(reachableLocations, inventory);
-                if (reachableLocations.Count == 0)
-                    return;
-
-                ItemLocation location = RemoveRandom(reachableLocations);
-                locations.Remove(location);
+                inventory.AddItem(item);
                 output.Add(location.id, item.id);
+                //Main.Randomizer.Log($"Placing prog item {item.id} at: {location.id}");
+
+                reachableLocations = FindReachableLocations(locations, inventory);
             }
         }
 
-        /// <summary>
-        /// Without logic, place the last item at a random location
-        /// </summary>
         private void FillJunkItems(List<ItemLocation> locations, List<Item> items, Dictionary<string, string> output)
         {
             ShuffleList(items);
@@ -182,34 +181,8 @@ namespace BlasII.Randomizer.Items
                 Item item = RemoveLast(items);
 
                 output.Add(location.id, item.id);
+                //Main.Randomizer.Log($"Placing junk item {item.id} at: {location.id}");
             }
         }
-
-        /// <summary>
-        /// After shuffling the list of progression items, move certain items to the end to prevent failing seeds
-        /// </summary>        
-        private void MovePriorityItems(List<Item> progressionItems)
-        {
-            Item wallClimb = Main.Randomizer.Data.GetItem("AB44");
-            progressionItems.Remove(wallClimb);
-            progressionItems.Insert(0, wallClimb);
-        }
-
-        /// <summary>
-        /// Finds all locations that are no longer reachable and removes them
-        /// </summary>
-        private void RemoveUnreachableLocations(List<ItemLocation> locations, Blas2Inventory inventory)
-        {
-            for (int i = 0; i < locations.Count; i++)
-            {
-                if (inventory.Evaluate(locations[i].logic))
-                    continue;
-
-                locations.RemoveAt(i);
-                i--;
-            }
-        }
-
-        #endregion
     }
 }
